add async cache and refactor
