+++ cognitojwt/__init__.py
@@ -1,4 +1,3 @@
 name = "example_pkg"
 
-from .cognito import decode, CognitoJWTException
-from .cognito_async import decode_async, CognitoAsyncJWTException
+from .cognito import decode, decode_async, CognitoJWTException
